# optimize small strings