This PR adds a couple of loop-unrolling code to reduce the number of iterations when dealing with small strings:
```
benchmark                                       old ns/op     new ns/op     delta
BenchmarkHash32/fnv1a/optimized/strlen=4        6.05          4.26          -29.59%
BenchmarkHash32/fnv1a/optimized/strlen=11       7.98          7.66          -4.01%
BenchmarkHash32/fnv1a/optimized/strlen=36       22.9          22.0          -3.93%
BenchmarkHash32/fnv1a/optimized/strlen=1000     870           874           +0.46%
BenchmarkHash64/fnv1a/optimized/strlen=4        6.42          4.88          -23.99%
BenchmarkHash64/fnv1a/optimized/strlen=11       8.44          7.51          -11.02%
BenchmarkHash64/fnv1a/optimized/strlen=36       23.5          23.3          -0.85%
BenchmarkHash64/fnv1a/optimized/strlen=1000     896           887           -1.00%

benchmark                                       old MB/s     new MB/s     speedup
BenchmarkHash32/fnv1a/optimized/strlen=4        660.81       938.89       1.42x
BenchmarkHash32/fnv1a/optimized/strlen=11       1378.41      1435.13      1.04x
BenchmarkHash32/fnv1a/optimized/strlen=36       1575.06      1633.63      1.04x
BenchmarkHash32/fnv1a/optimized/strlen=1000     1149.21      1143.00      0.99x
BenchmarkHash64/fnv1a/optimized/strlen=4        622.81       820.06       1.32x
BenchmarkHash64/fnv1a/optimized/strlen=11       1303.70      1463.94      1.12x
BenchmarkHash64/fnv1a/optimized/strlen=36       1533.61      1547.72      1.01x
BenchmarkHash64/fnv1a/optimized/strlen=1000     1114.86      1127.09      1.01x
```